// twilioflex/flex-template-agent-assist/src/types/feature-loader/FlexEvent.ts
/**
 * Copyright 2025 Google LLC
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *     http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

export enum FlexEvent {
  taskReceived = 'taskReceived',
  taskUpdated = 'taskUpdated',
  taskAccepted = 'taskAccepted',
  taskCanceled = 'taskCanceled',
  taskCompleted = 'taskCompleted',
  taskRejected = 'taskRejected',
  taskRescinded = 'taskRescinded',
  taskTimeout = 'taskTimeout',
  taskWrapup = 'taskWrapup',
  workerActivityUpdated = 'workerActivityUpdated',
  workerAttributesUpdated = 'workerAttributesUpdated',
  pluginsInitialized = 'pluginsInitialized',
  tokenUpdated = 'tokenUpdated',
}